Moved TaskUtils to tests
